+++ packages/experience/src/components/IdentifierRegisterForm/index.module.scss
